refactor: remove explict definition of fields in policy serializer
Initially, the approach is to explictly specify which information get
displayed in an API response.

However, the usage of `'__all__'` allow us to quickly work with updates to
schema, without having to come here.

However there is a security flaw catch as this would display all fields
as it is in the database, instead of a more formatted response.

A long term solution would be provided as soon as I can think of
something, probably we  would be needing another limited view of
`PolicySerializer` (just like `PolicyPurchaseResponseSerializer`).