feat: management command to purchase insurance policy
This patch aims to test out the domain-logic behind the policy purchase
algorithm of the policy service and determine the root cause of errors
of PR #74.

At hindsight, it would seem the issue with the said patch is due to
nested validation of the customer information, payment information,
policy activation information and other data within.

However, this established structure is need to enforce what data is
captred from the merchant - about a policy and the customer who is
buying it.