fix: change quote_data to quote_serializer
Previously the implementation tends to seek quote_data from the
`_get_quote_by_code` method which returns a serializer as opposed
to the assumed thought of, `serializer.data`.

Therefore, what is being accesssed was a serializer object rather than
a dict of serializedd data for the said quote.

This pattch resolves that by correcting the references where due.