refactor: implement notification channel methods
An error was trigger after policy purchase, and the application was
unable to send notification to the participant of the policy purchase
process - namely the merchant and the customer.

This error was caused by the fact that the PolicyNotificationService is
attempting to instantiate an abstract class EmailNotification with
abstract methods: `prepare_message` and `stream_through` which were not
defined on the `EmailNotification` Channel as their not within the
bounds of the channels domain.

To fix this, this patch introduces an approach of seperating the
Channels logic into a `NotificationChannel` interface - enforcing the
`prepare_message` and `send` methods for derived classes of the
interface.