test(Quotes Factory): data error - numberic field overflow
An error is thrown when attempting to create test data for Quotes with
the `QuoteFactory`. This is as a result of performing
`fake.random_number()` as a way to generate fake money values.

Faker however provide API to generate decimal values on the fly with
`pydecimal` method.

This patch corrects the said mistakes